connect users API to DDB helpers and setup types correctly
