+++ types/item.ts
@@ -1,5 +1,4 @@
 export interface Item {
   id: string;
-  uri?: string;
   displayName?: string;
 }
